python/requirements.txt
ctransformers
codetiming
langchain
langchain-community
openai
faiss-cpu
tiktoken
gradio
sentence-transformers
sentencepiece
protobuf==3.20.1
pypdf
pymupdf
pymupdf4llm
cryptography==42.0.5
transformers
einops
ninja
jiwer
FlagEmbedding
peft
biopython